backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Samir Rodas"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
